rename grizzly.environment to grizzly.behave
makes more sense, since it contains the behave "glue".
